feat: added driver asssit summary page hifi and connected export route to exporting the json
