+++ app/ui/src/app/driver_assist/styles.ts
@@ -1,0 +1,465 @@
+import type { CSSProperties } from "react";
+
+export const styles: Record<string, CSSProperties> = {
+  safeArea: {
+    minHeight: "100dvh",
+    backgroundColor: "#ffffff",
+    color: "#222222",
+    fontFamily: "var(--font-geist-sans), Arial, sans-serif",
+    paddingBottom: "calc(75px + env(safe-area-inset-bottom))",
+  },
+  loadingScreen: {
+    minHeight: "100dvh",
+    backgroundColor: "#ffffff",
+  },
+  uploadScreen: {
+    minHeight: "100dvh",
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
+    padding: 24,
+    backgroundColor: "#ffffff",
+  },
+  hiddenInput: {
+    display: "none",
+  },
+  appHeader: {
+    fontSize: 14,
+    fontWeight: 700,
+    color: "#202020",
+    margin: 0,
+  },
+  uploadButton: {
+    backgroundColor: "#111827",
+    border: 0,
+    borderRadius: 8,
+    color: "#ffffff",
+    cursor: "pointer",
+    fontSize: 16,
+    fontWeight: 600,
+    padding: "14px 20px",
+  },
+  errorText: {
+    color: "#b91c1c",
+    fontSize: 14,
+    marginTop: 14,
+    maxWidth: 300,
+    textAlign: "center",
+  },
+  container: {
+    width: "100%",
+    maxWidth: "none",
+    margin: "0 auto",
+    padding: "24px 12px 22px",
+  },
+  topBar: {
+    alignItems: "center",
+    display: "flex",
+    justifyContent: "space-between",
+    marginBottom: 20,
+    width: "100%",
+  },
+  iconButton: {
+    background: "transparent",
+    border: 0,
+    color: "#202020",
+    cursor: "pointer",
+    display: "inline-flex",
+    padding: 0,
+  },
+  summaryCard: {
+    marginBottom: 16,
+    width: "100%",
+  },
+  progressTrack: {
+    backgroundColor: "#dfdfde",
+    borderRadius: 999,
+    height: 6,
+    overflow: "hidden",
+  },
+  progressFill: {
+    backgroundColor: "#73b99f",
+    height: "100%",
+  },
+  statsRow: {
+    alignItems: "center",
+    alignSelf: "stretch",
+    display: "flex",
+    gap: 16,
+    justifyContent: "center",
+    marginBottom: 21,
+  },
+  statBlock: {
+    alignItems: "center",
+    background: "transparent",
+    border: 0,
+    cursor: "pointer",
+    display: "flex",
+    flex: "0 0 auto",
+    flexDirection: "column",
+    font: "inherit",
+    padding: 0,
+  },
+  statNumber: {
+    color: "#202020",
+    fontSize: 15,
+    fontWeight: 700,
+    lineHeight: 1,
+  },
+  statLabel: {
+    color: "#555555",
+    fontSize: 12,
+    marginTop: 7,
+  },
+  sectionTitle: {
+    color: "#202020",
+    fontSize: 13,
+    fontWeight: 700,
+    margin: "0 0 7px",
+  },
+  routeSection: {
+    scrollMarginTop: 16,
+  },
+  card: {
+    alignItems: "flex-start",
+    alignSelf: "stretch",
+    backgroundColor: "#FDFDFC",
+    border: "1px solid #DCDBD8",
+    borderRadius: 8,
+    display: "flex",
+    flexDirection: "column",
+    gap: 20,
+    marginBottom: 10,
+    padding: 16,
+  },
+  completedCard: {
+    opacity: 0.68,
+  },
+  failedCard: {
+    backgroundColor: "#fff6f6",
+  },
+  cardButton: {
+    alignItems: "flex-start",
+    background: "transparent",
+    border: 0,
+    cursor: "pointer",
+    display: "block",
+    font: "inherit",
+    padding: 0,
+    textAlign: "left",
+    width: "100%",
+  },
+  textBlock: {
+    display: "flex",
+    flex: 1,
+    flexDirection: "column",
+    minWidth: 0,
+  },
+  stopMetaRow: {
+    alignItems: "center",
+    display: "flex",
+    gap: 6,
+    marginBottom: 9,
+  },
+  stopNumberBadge: {
+    alignItems: "center",
+    backgroundColor: "#D5F2E8",
+    borderRadius: 2,
+    color: "#285241",
+    display: "inline-flex",
+    fontSize: 10,
+    height: 18,
+    justifyContent: "center",
+    lineHeight: 1,
+    minWidth: 20,
+  },
+  stopWindow: {
+    color: "#464544",
+    fontFamily:
+      "var(--font-manrope), var(--font-geist-sans), Arial, sans-serif",
+    fontSize: 14,
+    fontStyle: "normal",
+    fontWeight: 600,
+    lineHeight: "normal",
+  },
+  addressText: {
+    color: "#222222",
+    fontSize: 12,
+    fontWeight: 700,
+    lineHeight: 1.35,
+    marginBottom: 10,
+    overflowWrap: "anywhere",
+  },
+  infoLine: {
+    alignItems: "flex-start",
+    color: "#444444",
+    display: "flex",
+    gap: 7,
+    marginTop: 4,
+  },
+  infoIcon: {
+    color: "#2f3432",
+    display: "inline-flex",
+    flex: "0 0 16px",
+  },
+  infoText: {
+    color: "#444444",
+    flex: 1,
+    fontSize: 12,
+    lineHeight: 1.35,
+    overflowWrap: "anywhere",
+  },
+  expandedSection: {
+    marginTop: 13,
+    width: "100%",
+  },
+  noteInput: {
+    backgroundColor: "#ffffff",
+    border: "1px solid #dddddd",
+    borderRadius: 5,
+    color: "#222222",
+    font: "inherit",
+    fontSize: 12,
+    marginBottom: 8,
+    minHeight: 58,
+    padding: 8,
+    resize: "vertical",
+    width: "100%",
+  },
+  statusText: {
+    color: "#374151",
+    fontSize: 14,
+    margin: "0 0 12px",
+  },
+  buttonRow: {
+    display: "grid",
+    gap: 8,
+    gridTemplateColumns: "repeat(2, minmax(0, 1fr))",
+  },
+  actionButton: {
+    alignItems: "center",
+    backgroundColor: "#ffffff",
+    border: "1px solid #eeeeee",
+    borderRadius: 8,
+    color: "#222222",
+    cursor: "pointer",
+    display: "inline-flex",
+    fontSize: 11,
+    fontWeight: 500,
+    gap: 6,
+    justifyContent: "center",
+    minHeight: 42,
+    padding: "0 8px",
+  },
+  primaryActionButton: {
+    alignItems: "center",
+    alignSelf: "stretch",
+    backgroundColor: "#4CB599",
+    border: 0,
+    borderRadius: 16,
+    color: "#143228",
+    cursor: "pointer",
+    display: "inline-flex",
+    fontSize: 11,
+    fontWeight: 500,
+    gap: 6,
+    justifyContent: "center",
+    marginBottom: 7,
+    padding: "20px 24px",
+    width: "100%",
+  },
+  deliveredButton: {
+    alignItems: "center",
+    backgroundColor: "#ffffff",
+    border: "1px solid #222222",
+    borderRadius: 16,
+    color: "#222222",
+    cursor: "pointer",
+    display: "inline-flex",
+    fontSize: 11,
+    fontWeight: 500,
+    gap: 6,
+    justifyContent: "center",
+    marginBottom: 7,
+    padding: "20px 24px",
+    width: "100%",
+  },
+  historySection: {
+    marginTop: 29,
+    scrollMarginTop: 16,
+  },
+  historyTitle: {
+    color: "#202020",
+    fontSize: 13,
+    fontWeight: 700,
+    margin: "0 0 24px",
+  },
+  footer: {
+    marginTop: 36,
+    paddingBottom: 48,
+  },
+  footerLogo: {
+    display: "block",
+    height: 36,
+    objectFit: "contain",
+    width: 32,
+  },
+  footerText: {
+    color: "#202020",
+    fontSize: 16,
+    lineHeight: 1.45,
+    margin: 0,
+  },
+  finishBar: {
+    backgroundColor: "#ffffff",
+    border: "1px solid #dcdcdc",
+    borderBottom: 0,
+    borderTopLeftRadius: 6,
+    borderTopRightRadius: 6,
+    bottom: 0,
+    left: "50%",
+    maxWidth: "none",
+    padding: "10px 12px calc(10px + env(safe-area-inset-bottom))",
+    position: "fixed",
+    transform: "translateX(-50%)",
+    width: "100%",
+  },
+  finishButton: {
+    backgroundColor: "#4CB599",
+    border: 0,
+    borderRadius: 999,
+    color: "#143228",
+    cursor: "pointer",
+    fontSize: 14,
+    fontWeight: 500,
+    height: 41,
+    width: "100%",
+  },
+  modalBackdrop: {
+    alignItems: "center",
+    backgroundColor: "rgba(0, 0, 0, 0.2)",
+    bottom: 0,
+    display: "flex",
+    justifyContent: "center",
+    left: 0,
+    padding: 12,
+    position: "fixed",
+    right: 0,
+    top: 0,
+    zIndex: 50,
+  },
+  reportDialog: {
+    alignItems: "flex-start",
+    backgroundColor: "#FDFDFC",
+    borderRadius: 4,
+    boxShadow: "0 8px 24px rgba(0, 0, 0, 0.12)",
+    display: "flex",
+    flexDirection: "column",
+    gap: 8,
+    padding: 24,
+    width: 352,
+    maxWidth: "calc(100vw - 24px)",
+  },
+  reportHeader: {
+    alignItems: "center",
+    display: "flex",
+    justifyContent: "space-between",
+    width: "100%",
+  },
+  reportTitle: {
+    color: "#202020",
+    fontSize: 16,
+    fontWeight: 700,
+    lineHeight: 1.4,
+    margin: 0,
+  },
+  reportCloseButton: {
+    alignItems: "center",
+    background: "transparent",
+    border: 0,
+    color: "#202020",
+    cursor: "pointer",
+    display: "inline-flex",
+    fontSize: 24,
+    height: 28,
+    justifyContent: "center",
+    lineHeight: 1,
+    padding: 0,
+    width: 28,
+  },
+  reportPrompt: {
+    color: "#464544",
+    fontSize: 12,
+    lineHeight: 1.4,
+    margin: "8px 0 0",
+  },
+  required: {
+    color: "#C2410C",
+  },
+  reportOptions: {
+    display: "flex",
+    flexDirection: "column",
+    gap: 8,
+    width: "100%",
+  },
+  reportOption: {
+    alignItems: "center",
+    border: "1px solid #DCDBD8",
+    borderRadius: 4,
+    color: "#464544",
+    cursor: "pointer",
+    display: "flex",
+    fontSize: 12,
+    gap: 8,
+    minHeight: 32,
+    padding: "6px 8px",
+    width: "100%",
+  },
+  reportRadio: {
+    accentColor: "#464544",
+    height: 16,
+    margin: 0,
+    width: 16,
+  },
+  reportDetails: {
+    border: "1px solid #DCDBD8",
+    borderRadius: 4,
+    color: "#202020",
+    font: "inherit",
+    fontSize: 12,
+    minHeight: 72,
+    padding: 8,
+    resize: "vertical",
+    width: "100%",
+  },
+  reportActions: {
+    alignItems: "center",
+    alignSelf: "stretch",
+    display: "flex",
+    gap: 8,
+    justifyContent: "flex-end",
+    marginTop: 4,
+  },
+  reportCancelButton: {
+    backgroundColor: "transparent",
+    border: 0,
+    color: "#202020",
+    cursor: "pointer",
+    fontSize: 12,
+    fontWeight: 500,
+    minHeight: 36,
+    padding: "0 12px",
+  },
+  reportSubmitButton: {
+    backgroundColor: "#4CB599",
+    border: 0,
+    borderRadius: 4,
+    color: "#143228",
+    cursor: "pointer",
+    fontSize: 12,
+    fontWeight: 500,
+    minHeight: 36,
+    padding: "0 16px",
+  },
+};
