+++ app/ui/src/app/driver_assist/summary/styles.ts
@@ -1,0 +1,165 @@
+import type { CSSProperties } from "react";
+
+export const summaryStyles: Record<string, CSSProperties> = {
+  safeArea: {
+    backgroundColor: "#ffffff",
+    color: "#222222",
+    fontFamily: "var(--font-geist-sans), Arial, sans-serif",
+    minHeight: "100dvh",
+    padding: 0,
+  },
+  loadingScreen: {
+    backgroundColor: "#ffffff",
+    minHeight: "100dvh",
+  },
+  container: {
+    backgroundColor: "#ffffff",
+    margin: "0 auto",
+    minHeight: "100dvh",
+    padding: "40px 21px 156px",
+    width: "100%",
+  },
+  topBar: {
+    alignItems: "center",
+    display: "flex",
+    justifyContent: "space-between",
+    marginBottom: 34,
+  },
+  appHeader: {
+    color: "#252525",
+    fontSize: 20,
+    fontWeight: 700,
+    lineHeight: 1.2,
+    margin: 0,
+  },
+  warningButton: {
+    alignItems: "center",
+    background: "transparent",
+    border: 0,
+    color: "#4b4b4b",
+    cursor: "pointer",
+    display: "inline-flex",
+    padding: 0,
+  },
+  summarySection: {
+    marginBottom: 24,
+  },
+  sectionTitle: {
+    color: "#252525",
+    fontSize: 16,
+    fontWeight: 700,
+    lineHeight: 1.2,
+    margin: "0 0 14px",
+  },
+  statsRow: {
+    display: "grid",
+    gridTemplateColumns: "repeat(3, 1fr)",
+    margin: "0 auto",
+    maxWidth: 245,
+  },
+  statBlock: {
+    alignItems: "center",
+    display: "flex",
+    flexDirection: "column",
+    gap: 7,
+    minWidth: 0,
+  },
+  statNumber: {
+    color: "#252525",
+    fontSize: 20,
+    fontWeight: 700,
+    lineHeight: 1,
+  },
+  statLabel: {
+    color: "#4d4d4d",
+    fontSize: 16,
+    lineHeight: 1.2,
+  },
+  stopList: {
+    display: "flex",
+    flexDirection: "column",
+    gap: 16,
+  },
+  card: {
+    alignItems: "center",
+    backgroundColor: "#ffffff",
+    border: "1px solid #dedede",
+    borderRadius: 6,
+    display: "flex",
+    gap: 12,
+    justifyContent: "space-between",
+    minHeight: 129,
+    padding: "18px 16px",
+  },
+  stopText: {
+    minWidth: 0,
+  },
+  customerName: {
+    color: "#252525",
+    display: "block",
+    fontSize: 16,
+    fontWeight: 700,
+    lineHeight: 1.3,
+    marginBottom: 5,
+  },
+  address: {
+    color: "#333333",
+    fontSize: 16,
+    lineHeight: 1.45,
+    margin: "0 0 6px",
+    overflowWrap: "anywhere",
+  },
+  timestamp: {
+    color: "#666666",
+    fontSize: 14,
+    lineHeight: 1.35,
+    margin: 0,
+  },
+  statusBadge: {
+    backgroundColor: "#f7f7f7",
+    borderRadius: 5,
+    color: "#3f3f3f",
+    flex: "0 0 auto",
+    fontSize: 16,
+    lineHeight: 1.2,
+    padding: "6px 8px",
+  },
+  completeBadge: {
+    backgroundColor: "#dff0d8",
+    color: "#3e7a3b",
+  },
+  remainingBadge: {
+    backgroundColor: "#f7f7f7",
+    color: "#3f3f3f",
+  },
+  exportBar: {
+    backgroundColor: "#ffffff",
+    border: "1px solid #dcdcdc",
+    borderBottom: 0,
+    borderTopLeftRadius: 6,
+    borderTopRightRadius: 6,
+    bottom: 0,
+    left: 0,
+    padding: "10px 12px calc(10px + env(safe-area-inset-bottom))",
+    position: "fixed",
+    width: "100%",
+  },
+  exportButton: {
+    backgroundColor: "#4CB599",
+    border: 0,
+    borderRadius: 999,
+    color: "#143228",
+    cursor: "pointer",
+    fontSize: 14,
+    fontWeight: 500,
+    height: 41,
+    width: "100%",
+  },
+  exportMessage: {
+    color: "#3f3f3f",
+    fontSize: 12,
+    lineHeight: 1.3,
+    margin: "0 0 7px",
+    textAlign: "center",
+  },
+};
